+++ CHANGELOG.md
@@ -7,6 +7,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added 
+
+* `textarea` and related attributes (`cols` and `rows`)
+* `width_int` and `height_int` which takes an `i32` value instead of a string
+
 
 ## [1.0.0] - 2024-11-03